добавлен пример для торговли одновременно на двух брокерах по сигналам от RSI

+++ code_blocks/trading_robot_template_with_mt4/main.cpp
@@ -223,7 +223,7 @@ int main(int argc, char **argv) {
                         }
 
                         /* открываем сделку */
-                        uint32_t api_bet_id = 0;
+                        uint64_t api_bet_id = 0;
                         int err = intrade_bar_api.open_bo(
                             symbol_name,
                             "RSI",
@@ -291,7 +291,7 @@ int main(int argc, char **argv) {
                         }
 
                         /* открываем сделку */
-                        uint32_t api_bet_id = 0;
+                        uint64_t api_bet_id = 0;
                         int err = intrade_bar_api.open_bo(
                             symbol_name,
                             "RSI",
